+++ src/mainzone.js
@@ -24,8 +24,8 @@ class MainZone extends EventEmitter {
         this.name = device.name;
         this.zoneControl = device.zoneControl;
         this.inputsDisplayOrder = device.inputs?.displayOrder || 0;
-        this.buttons = (device.buttons || []).filter(button => (button.displayType ?? 0) > 0);
-        this.sensors = (device.sensors || []).filter(sensor => (sensor.displayType ?? 0) > 0 && sensor.mode >= 0);
+        this.buttons = (device.buttons ?? []).filter(button => (button.displayType ?? 0) > 0);
+        this.sensors = (device.sensors ?? []).filter(sensor => (sensor.displayType ?? 0) > 0 && (sensor.mode ?? -1) >= 0);
         this.powerControlZone = device.power?.zone || 0;
         this.volumeControl = device.volume?.displayType || 0;
         this.volumeControlZone = device.volume?.zone || 0;
@@ -1037,7 +1037,7 @@ class MainZone extends EventEmitter {
     async start() {
         try {
             //denon client
-            this.zone = new Zone(this.denon, this.device, this.inputsFile)
+            this.zone = new Zone(this.denon, this.device, this.inputsFile, this.restFul.enable, this.mqtt.enable)
                 .on('deviceInfo', (info) => {
                     this.emit('devInfo', `-------- ${this.name} --------`);
                     this.emit('devInfo', `Manufacturer: ${info.manufacturer}`);